First version of DNS compression
